requirements.txt

customtkinter==5.1.3
Pillow==10.0.1
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
pymediainfo==5.0.1
chardet==5.1.0
python-magic==0.4.27
